requirements.txt
Pillow==10.1.0
requests==2.31.0
streamlit==1.29.0

